requirements.txt
Django==4.0.3
django-ckeditor==6.2.0
Pillow==9.1.0
